+++ src/components/stepforms/schemas/substring.ts
@@ -1,4 +1,6 @@
-export default {
+import { addNotInColumnNamesConstraint, StepFormType } from './utils';
+
+const schema = {
   $schema: 'http://json-schema.org/draft-07/schema#',
   title: 'Substring step',
   type: 'object',
@@ -34,7 +36,17 @@ export default {
       },
       not: { const: 0 },
     },
+    newColumnName: {
+      type: 'string',
+      minLength: 1,
+      title: 'New column name',
+      description: 'The new column name to be created for the computation result',
+    },
   },
   required: ['column', 'start_index', 'end_index'],
   additionalProperties: false,
 };
+
+export default function buildSchema(form: StepFormType) {
+  return addNotInColumnNamesConstraint(schema, 'newColumnName', form.columnNames);
+}